"enable-pool-warning" command line flag added
If a pool or hosts of a pool are down for annoncement mining, then
there are many warning messages generated. This can confuse the user
and is very noisy. This commit removes these warnings, unless the
flag is specified. The goodrate percentage is sufficient to inform
the user that a pool is not working correctly. If the user wants to
see the details, they can enable the warnings with this flag.

Might be useful to allow this flag for announcement handler and block
miner as well. Currently in these other modes, the warnings are
still generated.